fix some dumb
