+++ code_samples/sharding/extended/index.js
