+++ code_samples/sharding/getting-started/bot.js
@@ -2,16 +2,6 @@ const Discord = require('discord.js');
 const client = new Discord.Client();
 const prefix = '!';
 
-function findEmoji(id) {
-	const temp = this.emojis.get(id);
-	if (!temp) return false;
-
-	const emoji = Object.assign({}, temp);
-	if (emoji.guild) emoji.guild = emoji.guild.id;
-	emoji.require_colons = emoji.requiresColons;
-
-	return emoji;
-}
 
 client.on('message', message => {
 	if (!message.content.startsWith(prefix) || message.author.bot) return;
@@ -25,38 +15,6 @@ client.on('message', message => {
 				return message.reply(`Server count: ${results.reduce((prev, val) => prev + val, 0)}`);
 			})
 			.catch(console.error);
-	}
-
-	if (command === 'send') {
-		if (!args.length) return message.reply('please specify a destination channel id.');
-
-		return client.shard.broadcastEval(`
-			const channel = this.channels.get('${args[0]}');
-			if (channel) {
-				channel.send('This is a message from shard ${this.shard.id}!');
-				true;
-			}
-			else false;
-		`)
-			.then(results => {
-				const found = results.find(result => result);
-				if (!found) return message.reply('I could not find such a channel.');
-
-				return message.reply(`I have sent a message to channel \`${args[0]}\`!`);
-			});
-	}
-
-	if (command === 'emoji') {
-		if (!args.length) return message.reply('please specify an emoji id to search for.');
-
-		return client.shard.broadcastEval(`(${findEmoji}).call(this, '${args[0]}')`)
-			.then(results => {
-				const found = results.find(result => result);
-				if (!found) return message.reply('I could not find such an emoji.');
-
-				const emoji = new Discord.Emoji(client.guilds.get(found.guild), found);
-				return message.reply(`I have found an emoji ${emoji}!`);
-			});
 	}
 });
